@@ -34,10 +34,17 @@
 	</profiles>
 
 	<properties>
-		<org.hibernate.version>4.3.8.Final</org.hibernate.version>
-		<org.hibernate.search.version>5.0.1.Final</org.hibernate.search.version>
-		<org.infinispan.version>7.1.1.Final</org.infinispan.version>
-		<com.amazonaws.version>1.9.0</com.amazonaws.version>
+		<java.version>1.8</java.version>
+		<spring-cloud-aws-context.verion>1.0.3.RELEASE</spring-cloud-aws-context.verion>
+		<hibernate.version>5.0.2.Final</hibernate.version>
+		<hibernate-search.version>5.5.0.Final</hibernate-search.version>
+		<infinispan-version>8.0.1.Final</infinispan-version>
+		<lucene.version>5.3.0</lucene.version>
+		<thymeleaf-extras-springsecurity4.version>2.1.2.RELEASE</thymeleaf-extras-springsecurity4.version>
+		<!-- Need 2.1.1
+		<thymeleaf-extras-java8time.version>2.1.0.RELEASE</thymeleaf-extras-java8time.version>
+		-->
+		<amazonaws-version>1.10.30</amazonaws-version>
 	</properties>
 
 	<repositories>
