@@ -20,7 +20,7 @@ import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.Type;
-import org.joda.time.LocalDateTime;
+import java.time.LocalDateTime;
 
 import javax.persistence.*;
 
@@ -44,7 +44,6 @@ public class PasswordResetToken extends DomainObject<String> {
 	private String email;
 
 	@Column(name = "expired_at", nullable = false)
-	@Type(type = "org.jadira.usertype.dateandtime.joda.PersistentLocalDateTime")
 	private LocalDateTime expiredAt;
 
 	@Override
