@@ -16,7 +16,7 @@
 
 package org.wallride.core.service;
 
-import org.joda.time.LocalDateTime;
+import java.time.LocalDateTime;
 
 import java.io.Serializable;
 import java.util.HashSet;
