@@ -64,8 +64,8 @@ public class SecurityConfig {
 
 	@Inject
 	private DataSource dataSource;
-	@Inject
-	private CacheManager cacheManager;
+//	@Inject
+//	private CacheManager cacheManager;
 
 	@Autowired
 	public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
@@ -125,6 +125,7 @@ public class SecurityConfig {
 					.and()
 				.headers()
 					.frameOptions().disable()
+					.and()
 				.csrf()
 					.disable()
 				.exceptionHandling()
@@ -203,6 +204,7 @@ public class SecurityConfig {
 					.and()
 				.headers()
 					.frameOptions().disable()
+					.and()
 				.csrf()
 					.disable()
 				.exceptionHandling()
@@ -227,7 +229,7 @@ public class SecurityConfig {
 
 	@Bean
 	public AffirmativeBased accessDecisionManager() {
-		List<AccessDecisionVoter> accessDecisionVoters = new ArrayList<>();
+		List<AccessDecisionVoter<?>> accessDecisionVoters = new ArrayList<>();
 		accessDecisionVoters.add(roleVoter());
 		accessDecisionVoters.add(webExpressionVoter());
 
