@@ -69,7 +69,7 @@ public class PostRepositoryImpl implements PostRepositoryCustom {
 					"title", "body",
 					"tags.name",
 			};
-			MultiFieldQueryParser parser = new MultiFieldQueryParser(Version.LATEST, fields, analyzer);
+			MultiFieldQueryParser parser = new MultiFieldQueryParser(fields, analyzer);
 			parser.setDefaultOperator(QueryParser.Operator.AND);
 			Query query = null;
 			try {
