@@ -66,7 +66,7 @@ public class UserRepositoryImpl implements UserRepositoryCustom {
 					"loginId",
 					"name.firstName", "name.lastName",
 			};
-			MultiFieldQueryParser parser = new MultiFieldQueryParser(Version.LATEST, fields, analyzer);
+			MultiFieldQueryParser parser = new MultiFieldQueryParser(fields, analyzer);
 			parser.setDefaultOperator(QueryParser.Operator.AND);
 			Query query = null;
 			try {
