@@ -20,15 +20,16 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
-import org.joda.time.LocalDateTime;
 
 import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
 
 public class LocalDateTimeSerializer extends JsonSerializer<LocalDateTime> {
 
 	@Override
 	public void serialize(LocalDateTime localDateTime, JsonGenerator jsonGenerator, SerializerProvider serializerProvider)
 			throws IOException, JsonProcessingException {
-		jsonGenerator.writeString(localDateTime.toString("yyyy/MM/dd HH:mm"));
+		jsonGenerator.writeString(localDateTime.format(DateTimeFormatter.ofPattern("yyyy/MM/dd HH:mm")));
 	}
 }
