@@ -16,13 +16,23 @@
 
 package org.wallride.web.support;
 
+import com.amazonaws.http.IdleConnectionReaper;
+import com.mysql.jdbc.AbandonedConnectionCleanupThread;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
 import javax.servlet.annotation.WebListener;
+import java.sql.Driver;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 import java.util.Enumeration;
 
 @WebListener
 public class WebListenerHandler implements ServletContextListener {
+
+	private final Logger logger = LoggerFactory.getLogger(WebListenerHandler.class);
 
 	@Override
 	public void contextInitialized(ServletContextEvent event) {
@@ -62,13 +72,37 @@ public class WebListenerHandler implements ServletContextListener {
 //				e.printStackTrace();
 //			}
 //		}
-//
-//		// MySQL driver leaves around a thread. This static method cleans it up.
-//		try {
-//			AbandonedConnectionCleanupThread.shutdown();
-//		}
-//		catch (InterruptedException e) {
-//			e.printStackTrace();
-//		}
+
+		try {
+			// http://stackoverflow.com/questions/18069042/spring-mvc-webapp-schedule-java-sdk-http-connection-reaper-failed-to-stop
+			IdleConnectionReaper.shutdown();
+
+			// MySQL driver leaves around a thread. This static method cleans it up.
+			AbandonedConnectionCleanupThread.shutdown();
+		}
+		catch (InterruptedException e) {
+			e.printStackTrace();
+		}
+
+		ClassLoader cl = Thread.currentThread().getContextClassLoader();
+
+		Enumeration<Driver> drivers = DriverManager.getDrivers();
+		while (drivers.hasMoreElements()) {
+			Driver driver = drivers.nextElement();
+
+			if (driver.getClass().getClassLoader() == cl) {
+
+				try {
+					logger.info("Deregistering JDBC driver {}", driver);
+					DriverManager.deregisterDriver(driver);
+
+				} catch (SQLException ex) {
+					logger.error("Error deregistering JDBC driver {}", driver, ex);
+				}
+
+			} else {
+				logger.trace("Not deregistering JDBC driver {} as it does not belong to this webapp's ClassLoader", driver);
+			}
+		}
 	}
 }
